+++ HISTORY.md
@@ -5,6 +5,8 @@
   functions.
 - Name resolution and overload resolution has been completely rewritten
   which should fix a number of long-standing issues.
+- Fixed a crash when a record aggregate contains an "others" association
+  and the fields have array types with different lengths.
 
 ## 1.5 - 2020-07-19
 - IEEE library sources are now distributed